Moved DTO validators into separate directories
